ai_report_agent/requirements.txt
# Multi-Agent Financial Report Generation System
# Python 3.10+

# LLM Frameworks
crewai>=0.80.0
langchain>=0.3.0
langchain-google-genai>=0.1.0
langgraph>=0.2.0

# Data Sources
yfinance>=0.2.40
tavily-python>=0.5.0
sec-edgar-downloader>=5.0.0

# Structured Output
pydantic>=2.7.0

# PDF Generation
reportlab>=4.2.0

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.10.0

# Utilities
python-dotenv>=1.0.1
rich>=13.7.0
tenacity>=8.3.0
streamlit>=1.28.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0


transformers>=4.30.0
torch>=2.0.0
pandas>=2.0.0
plotly>=5.0.0
